+++ rusage_unix.go
@@ -13,10 +13,10 @@ func init() {
 
 type rusageExtrasUnix struct{}
 
-func (x rusageExtrasUnix) maxRss(pst *os.ProcessState) (int64, bool) {
+func (x rusageExtrasUnix) maxRss(pst *os.ProcessState) (int, bool) {
 	rusage, ok := pst.SysUsage().(*syscall.Rusage)
 	if !ok {
 		return -1, false
 	}
-	return rusage.Maxrss, true
+	return int(rusage.Maxrss), true
 }
